refactor: 위도, 경도 자료형 수정

See merge request cloud_track/class_04/web_project3/team01/backend!40

+++ src/main/java/com/team01/billage/product/dto/ProductRequestDto.java
@@ -17,8 +17,8 @@ public class ProductRequestDto {
     private String description;
     private int dayPrice;
     private Integer weekPrice; // 선택값 (null 가능)
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private double latitude;
+    private double longitude;
     private List<ProductImageRequestDto> productImages;
 
 }
