+++ src/main/java/com/team01/billage/product/dto/ProductUpdateRequestDto.java
@@ -3,7 +3,6 @@ package com.team01.billage.product.dto;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -15,8 +14,8 @@ public class ProductUpdateRequestDto {
     private String description;
     private int dayPrice;
     private Integer weekPrice; // 선택값 (null 가능)
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private double latitude;
+    private double longitude;
     private List<ProductImageRequestDto> productImages; // 추가하는 이미지
     private List<ExistProductImageRequestDto> existProductImages; // 기존 이미지
